Tested with a multi thread config (`multi_thread.yaml`). Seems to work fine. Shortened default log messages published to the `sentor/event` topic. When executing actions using a simple action server sentor now provided feedback on the goal. Renamed config. Ros logs made during sentor initialisation are no longer published to the `sentor/event` topic. Updated pacakge.xml
To test with multi thread config simply launch the launch file `sentor.launch`.
As before send the robot in simulation to WayPoint1. The robot will automatically navigate to
WayPoint45. In the mean time sentor will execute a shell command `cowsay moo`. When the robot reaches its goal
it will teleport back to x=0,y=0 and relocalise.