Fixed an issue that was causing processes to be executed immediately (without taking `timeout` into account).
Previously, processes will be executed when the lambda conditions are satisfied. But they would not execute again unless they become unsatisfied, then satisfied again.
This is desirable behaviour in a lot of cases but maybe not all. So we now have the option to execute repeatedly (every timeout seconds), whilst the lambda conditions are satisfied.
See the new top level arg `oneshot` in the config.